test(Exporters): Fixed failing tests.
