+++ projects/igniteui-angular/src/lib/test-utils/grid-samples.spec.ts
@@ -2407,3 +2407,24 @@ export class GridExportGroupedDataComponent extends BasicGridComponent {
 export class MultiColumnHeadersExportComponent extends BasicGridComponent {
     public data = SampleTestData.contactInfoDataFull();
 }
+
+@Component({
+    template: `
+    <igx-grid #grid1 [data]="data">
+        <igx-column>
+            <ng-template igxCell>
+                <button>SimpleBtn</button>
+            </ng-template>
+        </igx-column>
+        <igx-column header="" field="ID"></igx-column>
+        <igx-column header="  " field=""></igx-column>
+        <igx-column header="Name" field="Name"></igx-column>
+        <igx-column header="JobTitle" field="JobTitle"></igx-column>
+    </igx-grid>`
+})
+
+export class GridWithEmptyColumnsComponent {
+    @ViewChild('grid1', { static: true }) public grid: IgxGridComponent;
+
+    public data = SampleTestData.personJobDataFull();
+}
